requirements.txt
gradio
pytesseract
Pillow
google-generativeai
